refactor: update architecture
