+++ src/app.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Home } from './components';
+import './components/theme/scss/index.scss';
 
 const App = () => <Home />;
