+++ src/components/index.ts
@@ -1,2 +1,1 @@
-export { CheckBox, Range, Button } from './shared';
 export { Home } from './Home';
