+++ src/components/theme/components/ThemeCheckBox/ThemeCheckBox.scss
