+++ src/components/theme/components/ThemeRange/ThemeRange.scss
