+++ src/components/theme/index.ts
@@ -1,0 +1,1 @@
+export { Button, Range, CheckBox } from './components';
